// src/shared/types/store.ts
export type OrderStatus = 'Pending' | 'Processing' | 'Shipped' | 'Delivered' | 'Cancelled' | 'Returned';
export type PaymentStatus = 'Paid' | 'Pending' | 'Refunded' | 'Failed';